+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [6.2.0-preview] - 2019-XX-XX
 
+### Fixed
+- Fixed post processing with stereo double-wide
+
 ## [6.1.0-preview] - 2019-02-13
 
 ### Added